Kek: don't count bot reacts
I'm surprised I didn't add this before. I'm doing this now because while bot-reacting doesn't change the saved data in the end, it does set the dirty flag, resulting in needless saving. It's a pittance compared to what else I do on the server's SSD, but I will get my money's worth out of this thing.